refactor based on Henry comments: add dual phase for execute and keep in memory
